+++ src/components/pages/main_page/Display.svelte
