+++ src/components/pages/main_page/user/NewUser.svelte
