+++ src/routes/(main)/[[kANDa]]/[[sarga]]/+layout.ts
